Remove unnecessary cast for setting Point object

+++ src/main/java/io/github/sacren/TestBall.java
@@ -5,7 +5,7 @@ public class TestBall {
     /** Ball test driver. */
     public static void main(String[] args) {
         /* Ball instance: individual values must be double. */
-        Ball ball = new Ball(new Point((double) 11, (double) 12), new Duo((double) 1, (double) 2));
+        Ball ball = new Ball(new Point(11, 12), new Duo((double) 1, (double) 2));
         System.out.println();
         System.out.println("=== Ball ===");
         System.out.println();
